requirements.txt
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
openinference-instrumentation==0.1.7